refactor: wearables-promise-utils.ts
